refactor: 레이아웃 분리
